Refactor and rewrite README
