+++ .gitignore
@@ -1,2 +1,4 @@
 .stack-work/
 *~
+elm-stuff
+dist
