+++ ChangeLog.md
@@ -1,5 +1,5 @@
 # Changelog for elm2nix
 
-## 0.0.1.0 (2017-12-14)
+## 0.0.1.0 (2018-12-28)
 
 - Initial release (Domen Kožar)
