+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright Domen Kožar (c) 2017
+Copyright Domen Kožar (c) 2017-2019
 
 All rights reserved.
