+++ data/default.nix
@@ -1,5 +1,6 @@
 { nixpkgs ? <nixpkgs>
-, config ? {} }:
+, config ? {}
+}:
 
 with (import nixpkgs config);
 
@@ -9,44 +10,28 @@ let
     , src
     , name
     , srcdir ? "./src"
-    , targets ? []
+    , targets ? ["Main"]
     }:
-    let
-      sources = import srcs { inherit fetchzip; };
-      exactDependencies = builtins.toFile "exact-dependencies.json"
-        (builtins.toJSON (lib.mapAttrs (name: value: value.version) sources));
-    in stdenv.mkDerivation {
+    stdenv.mkDerivation {
       inherit name src;
 
       buildInputs = [ elmPackages.elm ];
 
-      preConfigurePhases = ["setupElmStuffPhase"];
+      buildPhase = pkgs.elmPackages.fetchElmDeps {
+        elmPackages = import srcs;
+        versionsDat = ./versions.dat;
+      };
 
-      setupElmStuffPhase = ''
-        runHook preSetupElmStuffPhase
-
-        rm -rf elm-stuff
-        mkdir elm-stuff
-        ln -s \${exactDependencies} elm-stuff/exact-dependencies.json
-
-        \${lib.concatStrings (lib.mapAttrsToList (name: src: ''
-          mkdir -p elm-stuff/packages/\${name}
-          ln -s \${src.src} elm-stuff/packages/\${name}/\${src.version}
-        '') sources)}
-
-        runHook postSetupElmStuffPhase
-      '';
-
-      buildPhase = let
+      installPhase = let
         elmfile = module: "\${srcdir}/\${builtins.replaceStrings ["."] ["/"] module}.elm";
       in ''
         mkdir -p \$out/share/doc
         \${lib.concatStrings (map (module: ''
-          elm make \${elmfile module} --output \$out/\${module}.js --docs \$out/share/doc/\${module}.json
+          elm make \${elmfile module} \
+            --output \$out/\${module}.html \
+            --docs \$out/share/doc/\${module}.json
         '') targets)}
       '';
-
-      installPhase = ":";
     };
 in mkDerivation {
   name = "${name}";
