+++ elm2nix.cabal
@@ -10,11 +10,11 @@ homepage:       https://github.com/domenkozar/elm2nix#readme
 bug-reports:    https://github.com/domenkozar/elm2nix/issues
 author:         Domen Kožar
 maintainer:     domen@dev.si
-copyright:      2017 Domen Kožar
+copyright:      2017-2019 Domen Kožar
 license:        BSD3
 license-file:   LICENSE
 build-type:     Simple
-cabal-version:  >= 1.10
+cabal-version:  >= 2.0
 
 extra-source-files:
     ChangeLog.md
@@ -42,8 +42,8 @@ library
     , transformers
     , unordered-containers
   exposed-modules:
-      Lib
-      Prefetch
+      Elm2Nix
+      Elm2Nix.FixedOutput
   other-modules:
       Paths_elm2nix
   default-language: Haskell2010
