+++ scripts/tests.sh
@@ -1,11 +1,10 @@
 #!/usr/bin/env nix-shell
-#!nix-shell -I nixpkgs=https://github.com/NixOS/nixpkgs/archive/release-18.09.tar.gz -p git stack -i bash
+#!nix-shell -p git stack -i bash -I nixpkgs=https://github.com/NixOS/nixpkgs/archive/release-18.09.tar.gz
+export NIX_PATH=nixpkgs=https://github.com/NixOS/nixpkgs/archive/release-18.09.tar.gz
 
 set -e
 
-export NIX_PATH=nixpkgs=https://github.com/NixOS/nixpkgs/archive/release-18.09.tar.gz 
-
-stack install --nix
+stack install --nix --pedantic
 
 MYTMPDIR=$(mktemp -d)
 trap "rm -rf $MYTMPDIR" EXIT
@@ -23,5 +22,5 @@ pushd $MYTMPDIR/elm-todomvc
   ~/.local/bin/elm2nix init > default.nix
   ~/.local/bin/elm2nix convert > elm-srcs.nix
   nix-build
-  checkfile ./result/Material.js
+  checkfile ./result/Main.html
 popd
