+++ stack.yaml
@@ -1,4 +1,4 @@
-resolver: lts-12.20
+resolver: lts-13.0
 packages:
 - .
 nix:
